Reorg data sciecnce code
