+++ tests/test_data_science.py
@@ -1,3 +1,5 @@
+"""Tests for data science utilities."""
+
 import unittest
 
 import pandas as pd
@@ -6,9 +8,10 @@ import curies
 
 
 class TestDataScience(unittest.TestCase):
-    """"""
+    """Test case for data science utilities."""
 
     def test_case_mismatch(self):
+        """Test case mismatch on CURIE standardizations."""
         data = ["EFO:1", "nope:nope"]
         df = pd.DataFrame([(row,) for row in data], columns=["curie"])
 
@@ -18,3 +21,6 @@ class TestDataScience(unittest.TestCase):
 
         results = converter.pd_standardize_curie(df, column="curie")
         suggestions = results.get_suggestions()
+        self.assertIsInstance(suggestions, dict)
+        self.assertIn("", suggestions)
+        # FIXME add more detailed tests
